Move editorhtml/contentstate specific logic into submodules of rich_text
