+++ wagtail/documents/rich_text/contentstate.py
@@ -1,42 +1,7 @@
-from django.utils.html import escape
 from draftjs_exporter.dom import DOM
 
-from wagtail.admin.rich_text.converters import editor_html
 from wagtail.admin.rich_text.converters.html_to_contentstate import LinkElementHandler
 from wagtail.documents.models import get_document_model
-
-
-# Front-end conversion
-
-def document_linktype_handler(attrs):
-    Document = get_document_model()
-    try:
-        doc = Document.objects.get(id=attrs['id'])
-        return '<a href="%s">' % escape(doc.url)
-    except Document.DoesNotExist:
-        return "<a>"
-
-
-# hallo.js / editor-html conversion
-
-class DocumentLinkHandler:
-    @staticmethod
-    def get_db_attributes(tag):
-        return {'id': tag['data-id']}
-
-    @staticmethod
-    def expand_db_attributes(attrs):
-        Document = get_document_model()
-        try:
-            doc = Document.objects.get(id=attrs['id'])
-            return '<a data-linktype="document" data-id="%d" href="%s">' % (doc.id, escape(doc.url))
-        except Document.DoesNotExist:
-            return "<a>"
-
-
-EditorHTMLDocumentLinkConversionRule = [
-    editor_html.LinkTypeRule('document', DocumentLinkHandler),
-]
 
 
 # draft.js / contentstate conversion
